调整 Conclusions，支持公用
